+++ src/context.ts
@@ -1,5 +1,6 @@
 import { getCurrentInstance } from '@vue/composition-api'
-import { Context } from '@nuxt/types'
+
+import type { Context } from '@nuxt/types'
 
 interface ContextCallback {
   (context: Context): void
@@ -16,6 +17,20 @@ export const withContext = (callback: ContextCallback) => {
   callback(vm.$nuxt.context)
 }
 
+/**
+ * `useContext` which will return the Nuxt context.
+ * @example
+  ```ts
+  import { defineComponent, ref, useContext } from 'nuxt-composition-api'
+
+  export default defineComponent({
+    setup() {
+      const { store } = useContext()
+      store.dispatch('myAction')
+    },
+  })
+  ```
+ */
 export const useContext = () => {
   const vm = getCurrentInstance()
   if (!vm) throw new Error('This must be called within a setup function.')