+++ test/e2e/ssr-refs.ts
@@ -7,7 +7,7 @@ import {
 } from './helpers'
 
 // eslint-disable-next-line
-fixture`SSR Refs`
+fixture`ssrRef`
 
 test('Shows data on ssr-loaded page', async t => {
   await navigateTo('/ssr-ref')
@@ -21,12 +21,26 @@ test('Shows data on ssr-loaded page', async t => {
   await expectOnPage('ref-only SSR rendered')
 })
 
-// Needs further attention - data is being inlined in prod mode on `/`
-test.skip('Shows appropriate data on client-loaded page', async t => {
+test('Shows appropriate data on client-loaded page', async t => {
   await navigateTo('/')
   await t.click(Selector('a').withText('ssr refs'))
   await expectPathnameToBe('/ssr-ref')
   await expectNotOnPage('ref-only SSR rendered')
   await expectOnPage('function-runs SSR or client-side')
   await expectOnPage('on: client')
+})
+
+test('Shows SSR data when an ssrRef is defined outside setup', async () => {
+  await navigateTo('/no-setup')
+  await expectOnPage('ssrRef-SSR overwritten')
+  await expectOnPage('async-prefetched async')
+})
+
+test('Shows client-side only data when an ssrRef is defined outside setup', async t => {
+  await navigateTo('/')
+  await t.click(Selector('a').withText('outside of setup'))
+  await expectNotOnPage('ssrRef-SSR overwritten')
+  await expectNotOnPage('async-prefetched async')
+  await expectOnPage('ssrRef-default value')
+  await expectOnPage('async-default async')
 })