+++ test/fixture/pages/meta.vue
@@ -1,20 +1,26 @@
 <template>
   <div>
+    <div>title-{{ title }}</div>
     <nuxt-link to="/">link back</nuxt-link>
   </div>
 </template>
 
 <script>
-import { defineComponent, useHead, computed } from 'nuxt-composition-api'
-
-const { head, useMeta } = useHead()
+import { defineComponent, useMeta, computed } from 'nuxt-composition-api'
 
 export default defineComponent({
-  head,
+  head: {},
   setup() {
     const { title } = useMeta()
 
-    title.value = 'newSetTitle'
+    title.value = 'oldSetTitle'
+
+    const { title: newImport, bodyAttrs } = useMeta()
+    newImport.value = 'newSetTitle'
+
+    bodyAttrs.value.class = ['dark-mode', 'mobile']
+
+    return { title }
   },
 })
 </script>