+++ test/tsd/ssr-ref.test-d.ts
@@ -10,7 +10,6 @@ interface Obj {
 expectType<Ref<Obj>>(ssrRef({ name: 'today' }))
 expectType<Ref<null>>(ssrRef(null))
 
-
 expectType<Ref<number>>(shallowSsrRef(() => 42))
 expectType<Ref<string>>(shallowSsrRef('thoughtless'))
 expectType<Ref<Obj>>(shallowSsrRef({ name: 'today' }))