+++ test/unit/ssr-ref.spec.ts
@@ -1,6 +1,11 @@
 /**
  * @jest-environment jsdom
  */
+import Vue from 'vue'
+import CompositionApi from '@vue/composition-api'
+import { ssrRef, setSSRContext } from '../..'
+
+Vue.use(CompositionApi)
 
 jest.setTimeout(60000)
 
@@ -11,7 +16,7 @@ const config = require('../fixture/nuxt.config')
 
 let nuxt
 
-describe('SSR Refs', () => {
+describe('ssrRef', () => {
   beforeAll(async () => {
     nuxt = (await setup(config)).nuxt
   }, 60000)
@@ -21,8 +26,49 @@ describe('SSR Refs', () => {
   })
 
   test('__NUXT__ contains correct data', async () => {
-    const html = await get('/ssr-ref')
-    expect(html).toContain('"only SSR rendered"')
-    expect(html).toContain('"runs SSR or client-side"')
+    const homePage = await get('/')
+    expect(homePage.includes('"only SSR rendered"')).toBeFalsy()
+    expect(homePage.includes('"runs SSR or client-side"')).toBeFalsy()
+
+    const ssrRefPage = await get('/ssr-ref')
+    expect(ssrRefPage).toContain('"only SSR rendered"')
+    expect(ssrRefPage).toContain('"runs SSR or client-side"')
+
+    const noSetupPage = await get('/no-setup')
+    expect(noSetupPage).toContain('"prefetched async"')
+    expect(noSetupPage).toContain('"SSR overwritten"')
+    expect(noSetupPage.includes('"unchanged"')).toBeFalsy()
+
+    const rerenderedHomePage = await get('/')
+    expect(rerenderedHomePage.includes('"only SSR rendered"')).toBeFalsy()
+    expect(rerenderedHomePage.includes('"runs SSR or client-side"')).toBeFalsy()
+  })
+})
+
+describe('ssrRef reactivity', () => {
+  let ssrContext: Record<string, any>
+
+  beforeEach(async () => {
+    ssrContext = Object.assign({}, { nuxt: {} })
+    setSSRContext(ssrContext)
+  })
+  test('ssrRefs react to change in state', async () => {
+    process.client = false
+    const name = ssrRef('', 'name')
+    ssrRef('', 'unchanged')
+    name.value = 'full name'
+    expect(ssrContext).toMatchSnapshot()
+  })
+  test('ssrRefs react to deep change in object state', async () => {
+    process.client = false
+    const obj = ssrRef({ deep: { object: { name: 'nothing' } } }, 'obj')
+    obj.value.deep.object.name = 'full name'
+    expect(ssrContext).toMatchSnapshot()
+  })
+  test('ssrRefs react to deep change in array state', async () => {
+    process.client = false
+    const obj = ssrRef({ deep: { object: [{ name: 'nothing' }] } }, 'obj')
+    obj.value.deep.object[0].name = 'full name'
+    expect(ssrContext).toMatchSnapshot()
   })
 })